# pkg/latex/docs/swagger.yaml
basePath: /api/doc
info:
  contact:
    email: support@bullteam.cn
    name: API Support
  description: TEFS latex
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TEFS latex API
  version: V0.1
paths:
...
  /latex/{id}/pdf:
    post:
      consumes:
      - application/json
      description: 编译latex
      parameters:
      - description: latex ID
        in: path
        name: id
        required: true
        type: integer
      - description: latex 编译文件路径, 例如/main.tex(最外层目录)，/test/main.tex(test目录下文件)
        in: query
        name: filepath
        required: true
        type: string
      - description: 试试保存log文件
        in: query
        name: keep_logs
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 编译latex
      tags:
      - latex
  /latex/{id}/share:
    delete:
      consumes:
      - application/json
      description: 根据user_id和latex_id移除文档协作者
      parameters:
      - description: latex ID
        in: path
        name: id
        required: true
        type: integer
      - description: user ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 移除文档协作者
      tags:
      - share
    get:
      consumes:
      - application/json
      description: 根据latex_id查看当前latex文档协作者情况
      parameters:
      - description: latex ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 查询文档协作者列表
      tags:
      - share
    post:
      consumes:
      - application/json
      description: 邮件发送latex协作邀请URL
      parameters:
      - description: latex ID
        in: path
        name: id
        required: true
        type: integer
      - description: 目标用户邮箱
        in: query
        name: targetEmail
        required: true
        type: string
      - description: 权限：1=读写 2=只读
        in: query
        name: power
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 邮件发送latex协作邀请URL
      tags:
      - share
    put:
      consumes:
      - application/json
      description: 根据user_id和latex_id修改文档协作者权限
      parameters:
      - description: user_id
        in: query
        name: user_id
        required: true
        type: integer
      - description: latex_id
        in: path
        name: id
        required: true
        type: integer
      - description: 权限 1=读写 2=只读
        in: query
        name: power
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 修改文档协作者权限
      tags:
      - share
...
  /share:
    get:
      consumes:
      - application/json
      description: 根据latex_id查看当前latex文档邮件分享待处理情况
      parameters:
      - description: latex_id
        in: query
        name: id
        required: true
        type: integer
      - description: 从多少条开始
        in: query
        name: offset
        required: true
        type: integer
      - description: 返回的条数
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 查询文档分享URL待处理列表
      tags:
      - share
    post:
      consumes:
      - application/json
      description: 根据URL中的key获取latex进行协作
      parameters:
      - description: 分享latex的url中的key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 根据URL中的key获取latex进行协作
      tags:
      - share
  /share/{id}:
    delete:
      consumes:
      - application/json
      description: 根据share_id删除当前文档分享URL
      parameters:
      - description: share ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除文档分享URL，取消分享
      tags:
      - share
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"